LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/scrollIntoView-align-scrollport-covering-child-expected.txt


FAIL scrollIntoView scrolls scrollport-covering child in both axes assert_equals: center sets top expected 100 but got 0
